break out utils.mjs into smaller files inside of app dir
